Implemented and tested MoveAction. Refactored many adjacent places

+++ Assets/Scripts/Core/Models/GameProcess/ITurnPerformer.cs
@@ -5,7 +5,7 @@ namespace Core.Models.GameProcess
 	public interface ITurnPerformer
 	{
 		Faction Faction { get; }
-		IReactiveProperty<int> Oxygen { get; }
+		IReactiveProperty<short> Oxygen { get; }
 		IReactiveProperty<bool> HisTurn { get; }
 	}
 }
